Update CopyNameRow.vb
Major Change. Completely new architecture. Instead of copying all the rows/header etc. the master sheet is copied and all the non-filtered UID are deleted and file is saved as a new file. This way there is no need to state grouping location or file headers as the exact same file is saved minus all other UID